+++ CHANGELOG.md
@@ -5,3 +5,12 @@
 
 ### 1.0.1 (2024.11.17)
 - Added functionality to make variable image sizes dynamic and configurable via the plugin settings.
+
+### 1.0.2 (2024.12.05)
+- Added additional description text for the options in plugin settings
+
+### 1.0.3 (2024.12.05)
+- Added doc blocks for each function, also added additional prefix 'ubm'
+
+### 1.0.4 (2025.01.17)
+- Added additional functionality for the plugin settings page, bug fixes.